+++ Tests/TuistGeneratorTests/Generator/InfoPlistContentProviderTests.swift
@@ -17,7 +17,10 @@ final class InfoPlistContentProviderTests: XCTestCase {
         let target = Target.test(platform: .iOS, product: .app)
 
         // When
-        let got = subject.content(target: target, extendedWith: ["ExtraAttribute": "Value"])
+        let got = subject.content(graph: Graph.test(),
+                                  project: .empty(),
+                                  target: target,
+                                  extendedWith: ["ExtraAttribute": "Value"])
 
         // Then
         assertEqual(got, [
@@ -53,7 +56,10 @@ final class InfoPlistContentProviderTests: XCTestCase {
         let target = Target.test(platform: .macOS, product: .app)
 
         // When
-        let got = subject.content(target: target, extendedWith: ["ExtraAttribute": "Value"])
+        let got = subject.content(graph: Graph.test(),
+                                  project: .empty(),
+                                  target: target,
+                                  extendedWith: ["ExtraAttribute": "Value"])
 
         // Then
         assertEqual(got, [
@@ -79,7 +85,10 @@ final class InfoPlistContentProviderTests: XCTestCase {
         let target = Target.test(platform: .macOS, product: .framework)
 
         // When
-        let got = subject.content(target: target, extendedWith: ["ExtraAttribute": "Value"])
+        let got = subject.content(graph: Graph.test(),
+                                  project: .empty(),
+                                  target: target,
+                                  extendedWith: ["ExtraAttribute": "Value"])
 
         // Then
         assertEqual(got, [
@@ -101,7 +110,10 @@ final class InfoPlistContentProviderTests: XCTestCase {
         let target = Target.test(platform: .macOS, product: .staticLibrary)
 
         // When
-        let got = subject.content(target: target, extendedWith: ["ExtraAttribute": "Value"])
+        let got = subject.content(graph: Graph.test(),
+                                  project: .empty(),
+                                  target: target,
+                                  extendedWith: ["ExtraAttribute": "Value"])
 
         // Then
         XCTAssertNil(got)
@@ -112,20 +124,127 @@ final class InfoPlistContentProviderTests: XCTestCase {
         let target = Target.test(platform: .macOS, product: .dynamicLibrary)
 
         // When
-        let got = subject.content(target: target, extendedWith: ["ExtraAttribute": "Value"])
+        let got = subject.content(graph: Graph.test(),
+                                  project: .empty(),
+                                  target: target,
+                                  extendedWith: ["ExtraAttribute": "Value"])
 
         // Then
         XCTAssertNil(got)
     }
 
     func test_contentPackageType() {
-        assertPackageType(subject.content(target: .test(product: .app), extendedWith: [:]), "APPL")
-        assertPackageType(subject.content(target: .test(product: .unitTests), extendedWith: [:]), "BNDL")
-        assertPackageType(subject.content(target: .test(product: .uiTests), extendedWith: [:]), "BNDL")
-        assertPackageType(subject.content(target: .test(product: .bundle), extendedWith: [:]), "BNDL")
-        assertPackageType(subject.content(target: .test(product: .framework), extendedWith: [:]), "FMWK")
-        assertPackageType(subject.content(target: .test(product: .staticFramework), extendedWith: [:]), "FMWK")
+        func content(for target: Target) -> [String: Any]? {
+            subject.content(graph: Graph.test(),
+                            project: .empty(),
+                            target: target,
+                            extendedWith: [:])
+        }
+
+        assertPackageType(content(for: .test(product: .app)), "APPL")
+        assertPackageType(content(for: .test(product: .unitTests)), "BNDL")
+        assertPackageType(content(for: .test(product: .uiTests)), "BNDL")
+        assertPackageType(content(for: .test(product: .bundle)), "BNDL")
+        assertPackageType(content(for: .test(product: .framework)), "FMWK")
+        assertPackageType(content(for: .test(product: .staticFramework)), "FMWK")
+        assertPackageType(content(for: .test(product: .watch2App)), "$(PRODUCT_BUNDLE_PACKAGE_TYPE)")
     }
+
+    func test_content_whenWatchOSApp() {
+        // Given
+        let watchApp = Target.test(name: "MyWatchApp",
+                                   platform: .watchOS,
+                                   product: .watch2App)
+        let app = Target.test(platform: .iOS,
+                              product: .app,
+                              bundleId: "io.tuist.my.app.id")
+        let project = Project.test(targets: [
+            app,
+            watchApp,
+        ])
+        let graph = Graph.create(project: project, dependencies: [
+            (target: app, dependencies: [watchApp]),
+            (target: watchApp, dependencies: []),
+        ])
+
+        // When
+        let got = subject.content(graph: graph,
+                                  project: project,
+                                  target: watchApp,
+                                  extendedWith: [
+                                      "ExtraAttribute": "Value",
+                                  ])
+
+        // Then
+        assertEqual(got, [
+            "CFBundleName": "$(PRODUCT_NAME)",
+            "CFBundleShortVersionString": "1.0",
+            "CFBundlePackageType": "$(PRODUCT_BUNDLE_PACKAGE_TYPE)",
+            "UISupportedInterfaceOrientations": [
+                "UIInterfaceOrientationPortrait",
+                "UIInterfaceOrientationPortraitUpsideDown",
+            ],
+            "CFBundleIdentifier": "$(PRODUCT_BUNDLE_IDENTIFIER)",
+            "CFBundleInfoDictionaryVersion": "6.0",
+            "CFBundleVersion": "1",
+            "CFBundleDevelopmentRegion": "$(DEVELOPMENT_LANGUAGE)",
+            "CFBundleExecutable": "$(EXECUTABLE_NAME)",
+            "CFBundleDisplayName": "MyWatchApp",
+            "WKWatchKitApp": true,
+            "WKCompanionAppBundleIdentifier": "io.tuist.my.app.id",
+            "ExtraAttribute": "Value",
+
+        ])
+    }
+
+    func test_content_whenWatchOSAppExtension() {
+        // Given
+        let watchAppExtension = Target.test(name: "MyWatchAppExtension",
+                                            platform: .watchOS,
+                                            product: .watch2Extension)
+        let watchApp = Target.test(platform: .watchOS,
+                                   product: .watch2App,
+                                   bundleId: "io.tuist.my.app.id.mywatchapp")
+        let project = Project.test(targets: [
+            watchApp,
+            watchAppExtension,
+        ])
+        let graph = Graph.create(project: project, dependencies: [
+            (target: watchApp, dependencies: [watchAppExtension]),
+            (target: watchAppExtension, dependencies: []),
+        ])
+
+        // When
+        let got = subject.content(graph: graph,
+                                  project: project,
+                                  target: watchAppExtension,
+                                  extendedWith: [
+                                      "ExtraAttribute": "Value",
+                                  ])
+
+        // Then
+        assertEqual(got, [
+            "CFBundleName": "$(PRODUCT_NAME)",
+            "CFBundleShortVersionString": "1.0",
+            "CFBundlePackageType": "$(PRODUCT_BUNDLE_PACKAGE_TYPE)",
+            "CFBundleIdentifier": "$(PRODUCT_BUNDLE_IDENTIFIER)",
+            "CFBundleInfoDictionaryVersion": "6.0",
+            "CFBundleVersion": "1",
+            "CFBundleDevelopmentRegion": "$(DEVELOPMENT_LANGUAGE)",
+            "CFBundleExecutable": "$(EXECUTABLE_NAME)",
+            "CFBundleDisplayName": "MyWatchAppExtension",
+            "NSExtension": [
+                "NSExtensionAttributes": [
+                    "WKAppBundleIdentifier": "io.tuist.my.app.id.mywatchapp",
+                ],
+                "NSExtensionPointIdentifier": "com.apple.watchkit",
+            ],
+            "WKExtensionDelegateClassName": "$(PRODUCT_MODULE_NAME).ExtensionDelegate",
+            "ExtraAttribute": "Value",
+        ])
+    }
+
+    // MARK: - Helpers
 
     fileprivate func assertPackageType(_ lhs: [String: Any]?,
                                        _ packageType: String?,