+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Please, check out guidelines: https://keepachangelog.com/en/1.0.0/
 - Reactive interface to the System utility https://github.com/tuist/tuist/pull/770 by @pepibumur
 - Workflow to make sure that documentation and website build https://github.com/tuist/tuist/pull/783 by @pepibumur.
 - Support for `xcframework`  https://github.com/tuist/tuist/pull/769 by @lakpa
+- Support generating info.plist for Watch Apps & Extensions https://github.com/tuist/tuist/pull/756 by @kwridan
 
 ### Fixed